proXimidad/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-hosts==6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.15.4
mysqlclient==2.2.7
pillow==11.2.1
platformdirs==4.2.2
PyJWT==2.9.0
requests==2.31.0
setuptools==78.1.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.3
